requirements.txt
Package         Version
--------------- -------
appdirs         1.4.4
distlib         0.3.1
filelock        3.0.12
kafka-python    2.0.2
matplotlib	3.3.2
numpy           1.19.5
pandas          1.2.0
pip             20.3.3
pymongo         3.11.2
python-dateutil 2.8.1
pytz            2020.5
redis           3.5.3
setuptools      49.2.1
six             1.15.0
virtualenv      20.4.0
